Relocate webhook token verifier to new k8s.io/webhook-auth module (KEP-6060)
Per SIG-Auth review, move the webhook token verification code out of client-go
into a new, dependency-minimal staging module so webhooks (and controller-runtime)
can import it without pulling in the client-go dependency tree.

- Move verify/ (pure-stdlib core) and verify/admissionhttp/ (adapter) from
  staging/src/k8s.io/client-go/webhook/authentication/ to
  staging/src/k8s.io/webhook-auth/.
- Delete the go-jose-backed josekeyset package; a go-oidc-based KeySet will
  replace it in a follow-up. The adapter test now uses a stdlib fake KeySet.
- Add a minimal module go.mod (k8s.io/api is the only Kubernetes dependency)
  and an .import-restrictions forbidding k8s.io/client-go and k8s.io/kubernetes.

The module name is provisional pending SIG-Auth. The canonical go.mod/go.sum,
vendor wiring, and publishing rules must still be generated by the sanctioned
tooling (hack/update-vendor.sh); this commit establishes the code location to
anchor that discussion.
