+++ staging/src/k8s.io/webhook-auth/.import-restrictions
@@ -1,0 +1,10 @@
+rules:
+  # prevent import of k8s.io/kubernetes
+  - selectorRegexp: k8s[.]io/kubernetes
+    forbiddenPrefixes:
+      - ''
+  # prevent import of k8s.io/client-go — webhook-auth must stay below client-go
+  # so consumers can import it without pulling in the client-go dependency tree
+  - selectorRegexp: k8s[.]io/client-go
+    forbiddenPrefixes:
+      - ''
